Create dbo.usp_PandasMarketCalendarsPostImport.sql
Added stored procedure to move market calendar data from the dbimport.PandasMarketCalendars import table to the dbo.TradeDateCalendar target table, which is abstracted from the source of the market calendar base data.

This procedure will join the new data in the import table with existing data already in the target table, for a specific calendar id to ensure a coherent trade calendar. This updated calendar will then entirely replace the existing data in the target table for the specific calendar id.

Although this method of updating the target table trade calendar, it is required to ensure the NextBusDay and period offset calendars are updated correctly in the event of a change to the trade calendar.